Sync CrocSwap

+++ contracts/interfaces/ICrocLpConduit.sol
@@ -30,9 +30,9 @@ interface ICrocLpConduit {
      *           deposit. Reverts the transaction. */
     function depositCrocLiq (address sender, bytes32 poolHash,
                              int24 lowerTick, int24 upperTick,
-                             uint128 liq, uint64 mileage) external returns (bool);
+                             uint128 liq, uint72 mileage) external returns (bool);
 
     function withdrawCrocLiq (address sender, bytes32 poolHash,
                               int24 lowerTick, int24 upperTick,
-                              uint128 liq, uint64 mileage) external returns (bool);
+                              uint128 liq, uint72 mileage) external returns (bool);
 }
